requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Bcrypt
cryptography
Flask-Mail
Pillow
email_validator